Refactor server internals for better segmentation and maintainability
Broke down `server.go` monolith into multiple files for better maintainability. The `Application` struct, associated types, utility functions and hash functions have been extracted to individual files (`types.go`, `conv.go`). This allows us to follow Single Responsibility and separation of concerns principles better.

Also debug logging has been introduced for client repo insertion (`handlers.go`). This improves traceability of actions in the system which is crucial for identifying issues faster and monitoring the system's operation.

Existing functions have been moved without any changes to their workings. This ensures that no existing functionality is broken.
